requirements.txt
mcp>=0.5.0
fastapi>=0.95.0
uvicorn>=0.21.0
httpx>=0.24.0
python-dotenv>=1.0.1